requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-keyvault-certificates==4.8.0
azure-keyvault-keys==4.9.0
azure-keyvault-secrets==4.8.0
azure-mgmt-core==1.4.0
azure-mgmt-storage==21.2.1
azure-storage-blob==12.22.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
cryptography==43.0.1
debugpy==1.8.5
decorator==5.1.1
Django==5.1.1
executing==2.1.0
idna==3.8
ipykernel==6.29.5
ipython==8.27.0
isodate==0.6.1
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
msal==1.31.0
msal-extensions==1.2.0
nest-asyncio==1.6.0
nltk==3.9.1
packaging==24.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.2
portalocker==2.10.1
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
regex==2024.7.24
requests==2.32.3
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
